organize attributes parsing by types
