+++ bgpkit-parser/src/parser/bmp/messages/route_monitoring.rs
@@ -1,7 +1,6 @@
 use crate::parser::bgp::messages::parse_bgp_message;
 use crate::parser::bmp::error::ParserBmpError;
-use bgp_models::bgp::BgpMessage;
-use bgp_models::network::AsnLength;
+use bgp_models::prelude::*;
 
 #[derive(Debug)]
 pub struct RouteMonitoring {
