+++ bgpkit-parser/src/parser/mrt/messages/table_dump_message.rs
@@ -1,8 +1,7 @@
 use crate::error::*;
 use crate::parser::bgp::attributes::AttributeParser;
 use crate::parser::ReadUtils;
-use bgp_models::mrt::tabledump::TableDumpMessage;
-use bgp_models::network::{Afi, AsnLength, NetworkPrefix};
+use bgp_models::prelude::*;
 use byteorder::{ReadBytesExt, BE};
 use std::io::Cursor;
 use std::net::IpAddr;
